+++ packages/core/global-navigation/examples/00-default-global-navigation.js
@@ -1,0 +1,35 @@
+// @flow
+
+import React from 'react';
+import EmojiAtlassianIcon from '@atlaskit/icon/glyph/emoji/atlassian';
+import { LayoutManager, NavigationProvider } from '@atlaskit/navigation-next';
+
+import GlobalNavigation from '../src/components/GlobalNavigation';
+
+const Global = () => (
+  <GlobalNavigation
+    product={{
+      label: 'Jira',
+      icon: EmojiAtlassianIcon,
+    }}
+    search={{}}
+    create={{}}
+    people={{}}
+    notification={{}}
+    appSwitcher={{}}
+    help={{}}
+    profile={{}}
+  />
+);
+
+export default () => (
+  <NavigationProvider>
+    <LayoutManager
+      globalNavigation={Global}
+      productRootNavigation={() => null}
+      productContainerNavigation={() => null}
+    >
+      Page content
+    </LayoutManager>
+  </NavigationProvider>
+);