+++ packages/elements/analytics-listeners/__tests__/FabricAnalyticsListenersSpec.tsx
@@ -1,0 +1,44 @@
+import * as React from 'react';
+import { mount } from 'enzyme';
+import FabricAnalyticsListeners from '../src/FabricAnalyticsListeners';
+import {
+  DummyComponentWithAnalytics,
+  DummyComponent,
+} from '../example-helpers';
+import { AnalyticsWebClient } from '../src/types';
+
+describe('<FabricAnalyticsListeners />', () => {
+  let analyticsWebClientMock: AnalyticsWebClient;
+
+  beforeEach(() => {
+    analyticsWebClientMock = {
+      sendUIEvent: jest.fn(),
+      sendOperationalEvent: jest.fn(),
+      sendTrackEvent: jest.fn(),
+      sendScreenEvent: jest.fn(),
+    };
+  });
+
+  describe('<FabricElementsListener />', () => {
+    it('should listen and fire an UI event with analyticsWebClient', () => {
+      const compOnClick = jest.fn();
+      const component = mount(
+        <FabricAnalyticsListeners client={analyticsWebClientMock}>
+          <DummyComponentWithAnalytics onClick={compOnClick} />
+        </FabricAnalyticsListeners>,
+      );
+
+      const analyticsListener = component.find('FabricElementsListener');
+      expect(analyticsListener.props()).toHaveProperty(
+        'client',
+        analyticsWebClientMock,
+      );
+
+      const dummyComponent = analyticsListener.find(DummyComponent);
+      expect(dummyComponent).toHaveLength(1);
+
+      dummyComponent.simulate('click');
+      expect(compOnClick).toBeCalled();
+    });
+  });
+});