+++ packages/media/media-picker/src/components/__tests__/clipboard.spec.ts
@@ -3,8 +3,9 @@ import { MediaPickerContext } from '../../domain/context';
 import { UserEvent } from '../../outer/analytics/events';
 import { MockClipboardEvent, MockFile } from '../../util/clipboardEventMocks';
 import { Clipboard } from '../clipboard';
+import { UploadService } from '../../service/uploadServiceFactory';
 
-jest.mock('../../service/uploadService');
+jest.mock('../../service/uploadServiceFactory');
 
 class MockContext implements MediaPickerContext {
   trackEvent(event: UserEvent) {}
@@ -12,7 +13,7 @@ class MockContext implements MediaPickerContext {
 
 describe('Clipboard', () => {
   let clipboard: Clipboard;
-  let addFile: any;
+  let addFiles: any;
   const context = ContextFactory.create({
     serviceHost: '',
     authProvider: {} as any,
@@ -23,8 +24,8 @@ describe('Clipboard', () => {
     clipboard.activate();
     document.dispatchEvent(new Event('DOMContentLoaded'));
 
-    addFile = jest.fn();
-    (clipboard as any).uploadService.addFile = addFile;
+    addFiles = jest.fn();
+    ((clipboard as any).uploadService as UploadService).addFiles = addFiles;
 
     // necessary for dom ready listener
     setTimeout(done, 0);
@@ -34,26 +35,21 @@ describe('Clipboard', () => {
     clipboard.deactivate();
   });
 
-  it('should not call this.uploadService.addFile() when a paste event is dispatched without files', () => {
-    document.dispatchEvent(new MockClipboardEvent('paste'));
-    expect(addFile).toHaveBeenCalledTimes(0);
-  });
-
-  it('should call this.uploadService.addFile() when a paste event is dispatched with a single file', () => {
+  it('should call this.uploadService.addFiles() when a paste event is dispatched with a single file', () => {
     document.dispatchEvent(new MockClipboardEvent('paste', [new MockFile()]));
-    expect(addFile).toHaveBeenCalledTimes(1);
+    expect(addFiles).toHaveBeenCalledTimes(1);
   });
 
-  it('should call this.uploadService.addFile() when a paste event is dispatched with multiple files', () => {
-    document.dispatchEvent(
-      new MockClipboardEvent('paste', [new MockFile(), new MockFile()]),
-    );
-    expect(addFile).toHaveBeenCalledTimes(2);
+  it('should call this.uploadService.addFiles() when a paste event is dispatched with multiple files', () => {
+    const files = [new MockFile(), new MockFile()];
+    document.dispatchEvent(new MockClipboardEvent('paste', files));
+    expect(addFiles).toHaveBeenCalledTimes(1);
+    expect(addFiles).toHaveBeenCalledWith(files);
   });
 
-  it('should not call this.uploadService.addFile() when deactivated and a paste event is dispatched a single file', () => {
+  it('should not call this.uploadService.addFiles() when deactivated and a paste event is dispatched a single file', () => {
     clipboard.deactivate();
     document.dispatchEvent(new MockClipboardEvent('paste', [new MockFile()]));
-    expect(addFile).toHaveBeenCalledTimes(0);
+    expect(addFiles).toHaveBeenCalledTimes(0);
   });
 });