+++ website/src/index.js
@@ -1,6 +1,7 @@
 // @flow
 import React from 'react';
 import { render } from 'react-dom';
+
 // $FlowFixMe Required module not found
 import '@atlaskit/css-reset';
 import 'regenerator-runtime/runtime';